Created a single app controller
Now we can controll everything from one place because
both controllers are returned to the main controller.
Changed the structure od todoController, still have to check
if the functions work as intended.
From the console, you need to assign the returned
objects to variables to use them, otherwise the
objects created will be deleted and you can't use
functions on them.