Added tthe forms to modify todo details
Added forms inside of dialog tags to modify the todo
details. When clicking on cancel, the form resets to the
original todo details. When clicking on save, the todo
details are updated and the todo div is re-rendered.

Need to only dispaly the title, description and due date
(description and due date only when they exist).
The title is a must.
Need button to delete todos and projects.
Styling of the forms and todo div (change color based on priority).
Adding a storage mechanism to save todos and projects.