Form to create a new todo card
Todo card creation is based on a form now. The form
must be in a dialog otherwise it can't be closed properly.
I noticed that there are many dialogs showing in the Elements tab,
I need to optimize that later.
Need to implement the same for project creation.
Need the storage logic for todos and projects later.